use fuchsia clang to build llvm
This removed llvm's runtime dependencies on libc++, libgcc (statically linked to libc++, libc++abi, libunwind, compiler-rt) and speeds up llvm builds.
The fuchsia clang was chosen because it is built and tested by chromium luci with two-stages bootstrap LTO+PGO and from llvm main branch, so it has the latest optimizations and features of llvm main with guaranteed quality.